pyproject.toml
[project]
name = "ai-lunch-bot"
version = "1.0.0"
description = "すみよしランチメニューの問い合わせ・注文ができるローカルMCPサーバー"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0.0",
    "mcp[cli]>=1.0.0",
    "httpx>=0.27.0",
    "uvicorn>=0.30.0",
    "python-dotenv>=1.0.0",
    "beautifulsoup4>=4.12.0",
]

[project.scripts]
lunch-bot = "lunch_bot.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"
